+++ common/src/main/java/dev/boxadactle/boxlib/gui/auto/Slider.java
@@ -1,0 +1,14 @@
+package dev.boxadactle.boxlib.gui.auto;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Slider {
+
+    double min();
+    double max();
+
+    int decimalPlaces() default 2;
+
+}