+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
 - used arch
+- added auto config system